Simplify.
Except the move to a list comprehension, we drop str() -- it's only
purpose was to convert None (which we get from <a> tags without a
href, e.g. a local anchor) to a string, otherwise re.match would fail
later when trying to match(None). Instead we just drop None to begin with.