New handling of input.conf files
There is now an option for generating an input file. This so that you own
edits won't be removed. Also, removing the input file will generate the
default one. If you have your own edit and accidentally generate a new input,
you will have that backuped with an associated message.