Partial API implementation (broken)
The script is deactivated (`sys.exit()`) after the new reddit-API function
calls. I get "429 -- Too Many Requests"-errors all the time. Taking a look at
the API docs, that seems to be because I have not authenticated using OAuth2.

The rest of this commit body is regarding thoughts on how to get this to work.

The API-rules are found

 https://github.com/reddit-archive/reddit/wiki/API
 https://www.reddit.com/wiki/api

taking a look at

 https://github.com/reddit-archive/reddit/wiki/OAuth2

we notice that we are looking at an 'Installed app', so we can (*sigh*) make
this open source. This is the most annoying option, because they provide no
examples, and I have no experience.

Looking at

 https://github.com/reddit-archive/reddit/wiki/OAuth2-Quick-Start-Example

there are some quick python examples, but for 'web' and 'script' apps only.
But could be useful.

 https://praw.readthedocs.io/en/latest/getting_started/authentication.html#installed-application

uses PRAW, which seems to be some calling-the-reddit-api-python stuff, which I
*really* do not want to use. But, finally we have a little note one what to do
for 'installed' apps. It says

  The code flow can be used with an installed application just as described
 above with one change: set the value of client_secret to None when
 initializing Reddit.

so we could perhaps just not use the client_secret(?), but then why would
there be one?

Hopefully I will have some energy to do something about this. This is to
resolve Issue #8 once and for all. Hopefully.